python-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
docstring-parser==0.16
duckdb==1.2.1
fonttools==4.56.0
gitdb==4.0.12
gitpython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
googlesearch-python==1.3.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
lxml==5.3.1
markdown-it-py==3.0.0
matplotlib==3.10.1
mdurl==0.1.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pgvector==0.4.0
phidata==2.7.10
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.3
psycopg==3.2.6
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycountry==24.6.1
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.1
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.39
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing-extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
